refactor player: move playerTypes namespace from ResolutionState.h to Player.h
