+++ include/Minefield/player.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <vector>
+#include <Minefield/strongType.h>
 #include <Minefield/cell.h>
 struct Mine;
 struct Player
@@ -10,7 +11,18 @@ struct Player
     std::vector<Cell> disabledMineSpots; // locations where own mines were destroyed
     int remainingMines; // the number of mines the player can still place or use
 };
+namespace playerTypes
+{
+    struct AttackerTag{};
+    struct DefenderTag{};
+    struct Player1Tag{};
+    struct Player2Tag{};
 
+    using Attacker = StrongType<Player&, AttackerTag>;
+    using Defender = StrongType<Player&, DefenderTag>;
+    using Player1 = StrongType<Player&, Player1Tag>;
+    using Player2 = StrongType<Player&, Player2Tag>;
+}
 struct Mine
 {
     Cell location;
